backend/requirements.txt
Flask==2.3.2
flask_sqlalchemy==3.0.5
flask_bcrypt==1.0.1
flask_jwt_extended==4.4.4
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
python-dotenv==1.0.0
Flask-Cors==3.1.1
